+++ tests/parser/octave_bang/test_4.m
@@ -1,0 +1,3 @@
+function rv = test_4(x, !)
+    rv = x;
+end