+++ CHANGELOG.md
@@ -20,7 +20,8 @@ Not quite compatible with Octave yet. See #43 [octave support](https://github.co
 
 ### 0.9.29-dev
 
-
+* Support the `!` and `!=` operators in Octave mode. In Octave mode we
+  now also no longer parse `!` as if it was a MATLAB shell escape.
 
 ### 0.9.28